+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/preview-middleware
 
+## 1.0.17
+
+### Patch Changes
+
+- Updated dependencies [41b3908]
+    - @sap-ux/feature-toggle@1.0.2
+    - @sap-ux/adp-tooling@1.0.14
+    - @sap-ux/system-access@1.0.3
+
+## 1.0.16
+
+### Patch Changes
+
+- Updated dependencies [e66a8a9]
+    - @sap-ux/adp-tooling@1.0.13
+
 ## 1.0.15
 
 ### Patch Changes